api/requirements.txt
google-generativeai>=0.4.0
python-docx>=1.1.0
matplotlib>=3.8.0
pypdf>=4.0.0
pdfplumber>=0.10.0
Pillow>=10.0.0
